requirements.txt
Jinja2==2.11.2
mlrun==0.5.4
beautifulsoup4==4.9.3
PyYAML~=5.3.1
sphinx~=3.4.3
sphinx-book-theme==0.0.39
sphinx-copybutton==0.3.1
myst_nb==0.11.1
sphinx-rtd-theme==0.5.1